refactor(testing): share the v2 import resolver and compilation host across snapshot harnesses
