+++ crates/solidity-v2/outputs/cargo/tests/src/binder_output/runner.rs
@@ -2,37 +2,14 @@ use anyhow::{Result, ensure};
 use infra_utils::cargo::CargoWorkspace;
 use infra_utils::codegen::CodegenFileSystem;
 use infra_utils::paths::PathExtensions;
-use slang_solidity_v2::compilation::{CompilationBuilder, CompilationBuilderConfig, FileId};
+use slang_solidity_v2::compilation::FileId;
 use slang_solidity_v2_common::collections::SortedMap;
-use slang_solidity_v2_common::diagnostics::kinds::compilation::{MissingFile, UnresolvedImport};
+use solidity_v2_testing_utils::compilation;
 
 use super::report::binder_report;
 use super::report_data::ReportData;
 use crate::snapshots::{self, SnapshotOutcome, SnapshotStatus, TestConfig, TestMatrix};
 use crate::utils::multi_part_file::split_multi_file;
-use crate::utils::path_resolver;
-
-struct CompilationConfig {
-    files: SortedMap<FileId, String>,
-}
-
-impl CompilationBuilderConfig for CompilationConfig {
-    fn read_file(&mut self, file_id: &FileId) -> Result<String, MissingFile> {
-        self.files.get(file_id).cloned().ok_or_else(|| MissingFile {
-            reason: "File not found".to_string(),
-        })
-    }
-
-    fn resolve_import(
-        &mut self,
-        source_file_id: &FileId,
-        import_path: &str,
-    ) -> Result<FileId, UnresolvedImport> {
-        path_resolver::resolve_import(source_file_id, import_path).ok_or_else(|| UnresolvedImport {
-            reason: "Unresolved import".to_string(),
-        })
-    }
-}
 
 pub(crate) fn run(group_name: &str, test_name: &str) -> Result<()> {
     let test_dir = CargoWorkspace::locate_source_crate("solidity_v2_testing_snapshots")?
@@ -70,20 +47,7 @@ pub(crate) fn run(group_name: &str, test_name: &str) -> Result<()> {
         &test_config,
         "generated",
         |version, target| {
-            let compilation_config = CompilationConfig {
-                files: files.clone(),
-            };
-            let mut builder = CompilationBuilder::create(version, target, compilation_config);
-
-            // While `builder.add_file()` recursively adds dependencies, so adding
-            // the root file would be enough, we don't want to depend on the
-            // ordering of the parts in `input.sol`. Calling `add_file()` on files
-            // already added is idempotent, so to be sure we add all parts.
-            for file in files.keys() {
-                builder.add_file(file.clone());
-            }
-
-            let compilation = builder.build();
+            let compilation = compilation::compile(&files, version, target);
             let report_data = ReportData::prepare(&compilation, &files);
 
             let status = if report_data.all_resolved() {
