+++ crates/solidity-v2/outputs/cargo/tests/src/diagnostics_output/targets/slang.rs
@@ -1,14 +1,13 @@
 use anyhow::Result;
-use slang_solidity_v2::compilation::{CompilationBuilder, CompilationBuilderConfig, FileId};
+use slang_solidity_v2::compilation::FileId;
 use slang_solidity_v2_common::collections::SortedMap;
-use slang_solidity_v2_common::diagnostics::kinds::compilation::{MissingFile, UnresolvedImport};
 use slang_solidity_v2_common::diagnostics::{DiagnosticExtensions, DiagnosticSeverity};
 use slang_solidity_v2_common::evm_targets::EvmTarget;
 use slang_solidity_v2_common::versions::LanguageVersion;
+use solidity_v2_testing_utils::compilation;
 use solidity_v2_testing_utils::reporting::diagnostic;
 
 use crate::diagnostics_output::targets::{TargetOutcome, TestTarget};
-use crate::utils::path_resolver;
 
 pub(crate) struct SlangTarget;
 
@@ -23,16 +22,7 @@ impl TestTarget for SlangTarget {
         version: LanguageVersion,
         evm_target: EvmTarget,
     ) -> Result<TargetOutcome> {
-        let config = TestConfig {
-            files: files.clone(),
-        };
-        let mut builder = CompilationBuilder::create(version, evm_target, config);
-
-        for file in files.keys() {
-            builder.add_file(file.clone());
-        }
-
-        let compilation = builder.build();
+        let compilation = compilation::compile(files, version, evm_target);
 
         let diagnostics: Vec<_> = compilation.diagnostics().iter().collect();
 
@@ -53,28 +43,6 @@ impl TestTarget for SlangTarget {
         Ok(TargetOutcome {
             diagnostics: rendered,
             compilation_succeeded,
-        })
-    }
-}
-
-struct TestConfig {
-    files: SortedMap<FileId, String>,
-}
-
-impl CompilationBuilderConfig for TestConfig {
-    fn read_file(&mut self, file_id: &FileId) -> Result<String, MissingFile> {
-        self.files.get(file_id).cloned().ok_or_else(|| MissingFile {
-            reason: "File not found.".to_string(),
-        })
-    }
-
-    fn resolve_import(
-        &mut self,
-        source_file_id: &FileId,
-        import_path: &str,
-    ) -> Result<FileId, UnresolvedImport> {
-        path_resolver::resolve_import(source_file_id, import_path).ok_or_else(|| UnresolvedImport {
-            reason: "Unresolved import.".to_string(),
         })
     }
 }
