+++ crates/solidity-v2/outputs/cargo/tests/src/utils/mod.rs
@@ -1,2 +1,1 @@
 pub(crate) mod multi_part_file;
-pub(crate) mod path_resolver;
