+++ crates/solidity-v2/testing/utils/Cargo.toml
@@ -9,6 +9,7 @@ publish = false
 ariadne = { workspace = true }
 semver = { workspace = true }
 slang_solidity = { workspace = true }
+slang_solidity_v2 = { workspace = true }
 slang_solidity_v2_common = { workspace = true }
 slang_solidity_v2_cst = { workspace = true }
 slang_solidity_v2_parser = { workspace = true }
