+++ crates/solidity-v2/testing/utils/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod compilation;
 pub mod cst_renderer;
+pub mod path_resolver;
 pub mod reporting;
 pub mod v1_comparison;
