+++ crates/solidity-v2/testing/utils/src/path_resolver.rs
@@ -4,7 +4,7 @@ use slang_solidity_v2::compilation::FileId;
 ///
 /// `import_path` is the unquoted import path as provided by the v2
 /// `CompilationBuilderConfig::resolve_import` callback.
-pub(crate) fn resolve_import(source_file_id: &FileId, import_path: &str) -> Option<FileId> {
+pub fn resolve_import(source_file_id: &FileId, import_path: &str) -> Option<FileId> {
     if is_relative_path(import_path) {
         normalize_path(import_path, get_parent_path(source_file_id.as_str())).map(FileId::from)
     } else {
